fix(clippy): resolve all clippy warnings for Windows and cross-platform code
- Replace io::Error::new(ErrorKind::Other, ...) with io::Error::other(...)
  across console_windows, vsock_windows, whpx_vcpu, and builder
- Add Send/Sync impls for Registration (Windows HANDLE is thread-safe)
- Remove unnecessary type casts (u32->u32, u8->u8)
- Collapse nested if statements in net_windows
- Replace OR patterns with ranges (0xE4|0xE5|0xE6|0xE7 -> 0xE4..=0xE7)
- Remove redundant closure in net_windows config
- Fix field assignment outside initializer in builder
- Remove unused import round_up_to_page_size in rutabaga_gfx

All packages (utils, devices, vmm) now pass clippy with -D warnings.

+++ src/devices/src/virtio/fs/fuse.rs
@@ -610,7 +610,7 @@ impl Attr {
             gid: st.st_gid,
             #[cfg(target_os = "windows")]
             gid: st.st_gid as u32,
-            rdev: st.st_rdev as u32,
+            rdev: st.st_rdev,
             #[cfg(not(target_os = "windows"))]
             blksize: st.st_blksize as u32,
             #[cfg(target_os = "windows")]
